Remove DatabaseStructure JS global variable
Extracts the adjustTotals functions into a separate file.
